+++ packages/extension/src/state/test/update-structure.test.ts
@@ -1,8 +1,9 @@
-import { mapStructureUpdates, Structure } from "../structure"
+import { Structure } from "../structure"
+import { reconcileStructure } from "../structure-reconcile"
 import { describe, expect, test } from "vitest"
 import { Mapped, NodeID, NodeType } from "@solid-devtools/shared/graph"
 
-describe("mapStructureUpdates", () => {
+describe("reconcileStructure", () => {
   let updated: Record<NodeID, Mapped.Root> = {
     "1": {
       id: "1",
@@ -32,52 +33,24 @@ describe("mapStructureUpdates", () => {
       attachedTo: "3",
     },
   }
-
   let removed: NodeID[] = []
-  let prevStructure: Structure.Node[] = []
 
-  const attachments = new Map<NodeID, Mapped.Root[]>()
-  const mappedRoots = new Map<NodeID, Mapped.Root>()
+  let prevRoots: Structure.Node[] = []
 
   test("initial", () => {
-    const { structure } = mapStructureUpdates({
-      prev: [],
-      updated,
-      removed,
-      attachments,
-      mappedRoots,
-    })
-    prevStructure = structure
+    const { roots, nodeList } = reconcileStructure(prevRoots, updated, removed)
+    prevRoots = roots
 
-    expect(prevStructure, "initial structure").toMatchObject([
-      {
-        id: "1",
-        name: "root",
-        type: NodeType.Root,
-        length: 5,
-        children: [
-          { id: "2", name: "child", type: NodeType.Component, length: 0, children: [] },
-          {
-            id: "3",
-            name: "child2",
-            type: NodeType.Effect,
-            length: 3,
-            children: [
-              { id: "4", name: "child3", type: NodeType.Component, length: 0, children: [] },
-              {
-                id: "5",
-                name: "root2",
-                type: NodeType.Root,
-                length: 1,
-                children: [
-                  { id: "6", name: "child4", type: NodeType.Component, length: 0, children: [] },
-                ],
-              },
-            ],
-          },
-        ],
-      },
+    expect(nodeList).toMatchObject([
+      { id: "1", parent: null, children: [{ id: "2" }, { id: "3" }] },
+      { id: "2", parent: { id: "1" }, children: [] },
+      { id: "3", parent: { id: "1" }, children: [{ id: "4" }], subroots: [{ id: "5" }] },
+      { id: "4", parent: { id: "3" }, children: [] },
+      { id: "5", parent: { id: "3" }, children: [{ id: "6" }] },
+      { id: "6", parent: { id: "5" }, children: [] },
     ])
+
+    expect(roots).toMatchObject([{ id: "1" }])
   })
 
   test("update", () => {
@@ -89,6 +62,7 @@ describe("mapStructureUpdates", () => {
           name: "root",
           type: NodeType.Root,
           children: [
+            // removed child (2)
             {
               id: "3",
               name: "child2",
@@ -98,47 +72,32 @@ describe("mapStructureUpdates", () => {
           ],
         },
       },
+      // added new top-level root (10)
       "10": {
         id: "10",
         tree: { id: "10", name: "root4", type: NodeType.Root, children: [] },
       },
       "5": {
         id: "5",
+        // removed child (6)
         tree: { id: "5", name: "root2", type: NodeType.Root, children: [] },
         attachedTo: "3",
       },
     }
 
-    const { structure } = mapStructureUpdates({
-      prev: prevStructure,
-      updated,
-      removed,
-      attachments,
-      mappedRoots,
-    })
-    prevStructure = structure
+    const { roots, nodeList } = reconcileStructure(prevRoots, updated, removed)
+    prevRoots = roots
 
-    expect(prevStructure).toMatchObject([
-      {
-        id: "1",
-        name: "root",
-        type: NodeType.Root,
-        length: 3,
-        children: [
-          {
-            id: "3",
-            name: "child2",
-            type: NodeType.Effect,
-            length: 2,
-            children: [
-              { id: "4", name: "child3", type: NodeType.Component, length: 0, children: [] },
-              { id: "5", name: "root2", type: NodeType.Root, length: 0, children: [] },
-            ],
-          },
-        ],
-      },
-      { id: "10", name: "root4", type: NodeType.Root, length: 0, children: [] },
+    expect(nodeList).toHaveLength(5)
+    expect(nodeList).toMatchObject([
+      { id: "1", parent: null, children: [{ id: "3" }] },
+      { id: "3", parent: { id: "1" }, children: [{ id: "4" }], subroots: [{ id: "5" }] },
+      { id: "4", parent: { id: "3" }, children: [] },
+      { id: "5", parent: { id: "3" }, children: [] },
+      { id: "10", parent: null, children: [] },
     ])
+
+    expect(roots).toMatchObject([{ id: "1" }, { id: "10" }])
   })
 
   test("attach to a different node", () => {
@@ -146,17 +105,20 @@ describe("mapStructureUpdates", () => {
     updated = {
       "5": {
         id: "5",
+        // changed attachment from 3 to 8
         attachedTo: "8",
         tree: {
           id: "5",
           name: "root2",
           type: NodeType.Root,
           children: [
+            // add two children (6, 7)
             { id: "6", name: "child4", type: NodeType.Component, children: [] },
             { id: "7", name: "child5", type: NodeType.Refresh, children: [] },
           ],
         },
       },
+      // add new root (8)
       "8": {
         id: "8",
         tree: {
@@ -170,59 +132,31 @@ describe("mapStructureUpdates", () => {
 
     removed = ["10"]
 
-    const { structure } = mapStructureUpdates({
-      prev: prevStructure,
-      updated,
-      removed,
-      attachments,
-      mappedRoots,
-    })
-    prevStructure = structure
+    const { roots, nodeList } = reconcileStructure(prevRoots, updated, removed)
+    prevRoots = roots
 
-    expect(prevStructure, "updated structure 2").toMatchObject([
-      {
-        id: "1",
-        name: "root",
-        type: NodeType.Root,
-        length: 2,
-        children: [
-          {
-            id: "3",
-            name: "child2",
-            type: NodeType.Effect,
-            length: 1,
-            children: [
-              { id: "4", name: "child3", type: NodeType.Component, length: 0, children: [] },
-            ],
-          },
-        ],
-      },
-      {
-        id: "8",
-        name: "root3",
-        type: NodeType.Root,
-        length: 4,
-        children: [
-          { id: "9", name: "child6", type: NodeType.Component, length: 0, children: [] },
-          {
-            id: "5",
-            name: "root2",
-            type: NodeType.Root,
-            length: 2,
-            children: [
-              { id: "6", name: "child4", type: NodeType.Component, length: 0, children: [] },
-              { id: "7", name: "child5", type: NodeType.Refresh, length: 0, children: [] },
-            ],
-          },
-        ],
-      },
+    expect(nodeList).toHaveLength(8)
+    expect(nodeList).toMatchObject([
+      { id: "1", parent: null, children: [{ id: "3" }] },
+      { id: "3", parent: { id: "1" }, children: [{ id: "4" }] },
+      { id: "4", parent: { id: "3" }, children: [] },
+      { id: "8", parent: null, children: [{ id: "9" }], subroots: [{ id: "5" }] },
+      { id: "9", parent: { id: "8" }, children: [] },
+      { id: "5", parent: { id: "8" }, children: [{ id: "6" }, { id: "7" }] },
+      { id: "6", parent: { id: "5" }, children: [] },
+      { id: "7", parent: { id: "5" }, children: [] },
     ])
+
+    expect(roots).toHaveLength(2)
+    expect(roots[0].id).toBe("1")
+    expect(roots[1].id).toBe("8")
   })
 
   test("remove attachment", () => {
     removed = ["1"]
 
     updated = {
+      // remove attachment from 8
       "5": {
         id: "5",
         tree: {
@@ -235,11 +169,13 @@ describe("mapStructureUpdates", () => {
               id: "7",
               name: "child5",
               type: NodeType.Refresh,
+              // add a child (11)
               children: [{ id: "11", name: "child7", type: NodeType.Computation, children: [] }],
             },
           ],
         },
       },
+      // add new subroot (12) attached to 9
       "12": {
         id: "12",
         tree: { id: "12", name: "root5", type: NodeType.Root, children: [] },
@@ -247,161 +183,89 @@ describe("mapStructureUpdates", () => {
       },
     }
 
-    const { structure } = mapStructureUpdates({
-      prev: prevStructure,
-      updated,
-      removed,
-      attachments,
-      mappedRoots,
-    })
-    prevStructure = structure
+    const { roots, nodeList } = reconcileStructure(prevRoots, updated, removed)
+    prevRoots = roots
 
-    expect(prevStructure).toMatchObject([
-      {
-        id: "8",
-        name: "root3",
-        type: NodeType.Root,
-        length: 2,
-        children: [
-          {
-            id: "9",
-            name: "child6",
-            type: NodeType.Component,
-            length: 1,
-            children: [{ id: "12", name: "root5", type: NodeType.Root, length: 0, children: [] }],
-          },
-        ],
-      },
-      {
-        id: "5",
-        name: "root2",
-        type: NodeType.Root,
-        length: 3,
-        children: [
-          { id: "6", name: "child4", type: NodeType.Component, length: 0, children: [] },
-          {
-            id: "7",
-            name: "child5",
-            type: NodeType.Refresh,
-            length: 1,
-            children: [
-              { id: "11", name: "child7", type: NodeType.Computation, length: 0, children: [] },
-            ],
-          },
-        ],
-      },
+    expect(nodeList).toHaveLength(7)
+    expect(nodeList).toMatchObject([
+      { id: "8", parent: null, children: [{ id: "9" }] },
+      { id: "9", parent: { id: "8" }, children: [], subroots: [{ id: "12" }] },
+      { id: "12", parent: { id: "9" }, children: [] },
+      { id: "5", parent: null, children: [{ id: "6" }, { id: "7" }] },
+      { id: "6", parent: { id: "5" }, children: [] },
+      { id: "7", parent: { id: "5" }, children: [{ id: "11" }] },
+      { id: "11", parent: { id: "7" }, children: [] },
     ])
+
+    expect(roots).toHaveLength(2)
+    expect(roots[0].id).toBe("8")
+    expect(roots[1].id).toBe("5")
   })
 
   test("remove attached node", () => {
     removed = ["12"]
     updated = {}
 
-    const { structure } = mapStructureUpdates({
-      prev: prevStructure,
-      updated,
-      removed,
-      attachments,
-      mappedRoots,
-    })
-    prevStructure = structure
+    const { roots, nodeList } = reconcileStructure(prevRoots, updated, removed)
+    prevRoots = roots
 
-    expect(prevStructure).toMatchObject([
-      {
-        id: "8",
-        name: "root3",
-        type: NodeType.Root,
-        length: 1,
-        children: [{ id: "9", name: "child6", type: NodeType.Component, length: 0, children: [] }],
-      },
-      {
-        id: "5",
-        name: "root2",
-        type: NodeType.Root,
-        length: 3,
-        children: [
-          { id: "6", name: "child4", type: NodeType.Component, length: 0, children: [] },
-          {
-            id: "7",
-            name: "child5",
-            type: NodeType.Refresh,
-            length: 1,
-            children: [
-              { id: "11", name: "child7", type: NodeType.Computation, length: 0, children: [] },
-            ],
-          },
-        ],
-      },
+    expect(nodeList).toHaveLength(6)
+    expect(nodeList).toMatchObject([
+      { id: "8", parent: null, children: [{ id: "9" }] },
+      { id: "9", parent: { id: "8" }, children: [] },
+      { id: "5", parent: null, children: [{ id: "6" }, { id: "7" }] },
+      { id: "6", parent: { id: "5" }, children: [] },
+      { id: "7", parent: { id: "5" }, children: [{ id: "11" }] },
+      { id: "11", parent: { id: "7" }, children: [] },
     ])
+
+    expect(roots).toHaveLength(2)
+    expect(roots[0].id).toBe("8")
+    expect(roots[1].id).toBe("5")
   })
 })
 
 describe("mapArray example", () => {
   test("items keeping their position after update", () => {
-    let removed: NodeID[] = []
-    let updated: Record<NodeID, Mapped.Root> = {
-      "1": {
-        id: "1",
-        tree: {
-          id: "1",
-          name: "root",
-          type: NodeType.Root,
-          children: [{ id: "2", name: "For", type: NodeType.Component, children: [] }],
-        },
-      },
-      "3": {
-        id: "3",
-        tree: { id: "3", name: "item1", type: NodeType.Root, children: [] },
-        attachedTo: "2",
-      },
-      "4": {
-        id: "4",
-        tree: { id: "4", name: "item2", type: NodeType.Root, children: [] },
-        attachedTo: "2",
-      },
-      "5": {
-        id: "5",
-        tree: { id: "5", name: "item3", type: NodeType.Root, children: [] },
-        attachedTo: "2",
-      },
+    const n1: Structure.Node = {
+      id: "1",
+      name: "root",
+      type: NodeType.Root,
+      parent: null,
+      children: [],
     }
+    const n2: Structure.Node = {
+      id: "2",
+      name: "For",
+      type: NodeType.Component,
+      parent: n1,
+      children: [],
+    }
+    const n3: Structure.Node = {
+      id: "3",
+      name: "item1",
+      type: NodeType.Root,
+      parent: n2,
+      children: [],
+    }
+    const n4: Structure.Node = {
+      id: "4",
+      name: "item2",
+      type: NodeType.Root,
+      parent: n2,
+      children: [],
+    }
+    const n5: Structure.Node = {
+      id: "5",
+      name: "item3",
+      type: NodeType.Root,
+      parent: n2,
+      children: [],
+    }
+    ;(n1 as any).children = [n2]
+    ;(n2 as any).subroots = [n3, n4, n5]
 
-    const attachments = new Map<NodeID, Mapped.Root[]>()
-    const mappedRoots = new Map<NodeID, Mapped.Root>()
-
-    const { structure } = mapStructureUpdates({
-      prev: [],
-      updated,
-      removed,
-      attachments,
-      mappedRoots,
-    })
-
-    const goalStructure = [
-      {
-        id: "1",
-        name: "root",
-        type: NodeType.Root,
-        length: 4,
-        children: [
-          {
-            id: "2",
-            name: "For",
-            length: 3,
-            type: NodeType.Component,
-            children: [
-              { id: "3", name: "item1", length: 0, type: NodeType.Root, children: [] },
-              { id: "4", name: "item2", length: 0, type: NodeType.Root, children: [] },
-              { id: "5", name: "item3", length: 0, type: NodeType.Root, children: [] },
-            ],
-          },
-        ],
-      },
-    ]
-
-    expect(structure, "init").toMatchObject(goalStructure)
-
-    updated = {
+    const updated: Record<NodeID, Mapped.Root> = {
       "3": {
         id: "3",
         tree: { id: "3", name: "item1", type: NodeType.Root, children: [] },
@@ -409,14 +273,12 @@ describe("mapArray example", () => {
       },
     }
 
-    const { structure: structure1 } = mapStructureUpdates({
-      prev: structure,
-      updated,
-      removed,
-      attachments,
-      mappedRoots,
-    })
+    const prevRoots: Structure.Node[] = [n1]
+    const prevNodeList: Structure.Node[] = [n1, n2, n3, n4, n5]
 
-    expect(structure1, "after update").toMatchObject(goalStructure)
+    const { roots, nodeList } = reconcileStructure(prevRoots, updated, [])
+
+    expect(roots).toEqual(prevRoots)
+    expect(nodeList).toEqual(prevNodeList)
   })
 })
